+++ apps/users/serializers/admin_withdrawal_serializers.py
@@ -4,7 +4,7 @@ from rest_framework import serializers
 
 from apps.users.models import User
 from apps.users.models.withdrawal import Withdrawal
-from apps.users.utils.reason_choices import WithdrawalReason
+from apps.users.utils.conts import WithdrawalReason
 
 
 class AdminWithdrawalListItemSerializer(serializers.Serializer[Any]):
